add bandwidth gate simnet tests
